requirements.txt
streamlit
pandas
numpy
plotly
pymavlink
google-generativeai